beatmap timing improved
Instead of basing the next object on the last object(relative timing), the game will now base itself on a set "gametime", which is a universal timer which will sync everything up(universal timing). This will make longer maps stay in sync and allow people with worse devices to play in sync.